Chromoting Android app mouse/keyboard bugfixes and enhancements
- Enable multitouch panning on devices whose gesture detectors filture out such events by default
- Fix problem where the transformation matrix was applied twice when tapping to single click
- Hardcode the expansions for the @#*+ symbols for now
- Fix issue where hitting the keypad enter key could close the software keyboard without warning

Review URL: https://chromiumcodereview.appspot.com/21236002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@214769 0039d316-1c4b-4281-b951-d872f2087c98

+++ remoting/client/jni/android_keymap.h
@@ -11,13 +11,13 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 
 namespace remoting {
 
-// This is a temporary hack to convert Android's virtual keycodes into USB
-// scancodes. Unfortunately, the current solution uses key mappings whose
-// layout is determined according to the client device's input locale instead
-// of the host's internationalization preferences. The whole process needs to
-// be rethought to accomplish this in an localizable and future-proof way.
-// TODO(solb): crbug.com/265945
-uint32 AndroidKeycodeToUsbKeycode(int android);
+// TODO(solb): This is a temporary hack to convert Android's virtual keycodes
+// into USB scancodes. Unfortunately, the current solution uses key mappings
+// whose layout is determined according to the client device's input locale
+// instead of the host's. The whole process needs to be rethought to accomplish
+// the mappings in a localizable and future-proof way.
+// crbug.com/265945
+uint32 AndroidKeycodeToUsbKeycode(size_t android);
 
 }  // namespace remoting
